Moved PostgresConfig into internal/armada/configuration
