+++ internal/lookout/metrics/db_metrics_provider.go
@@ -3,7 +3,7 @@ package metrics
 import (
 	"database/sql"
 
-	"github.com/G-Research/armada/internal/lookout/configuration"
+	"github.com/G-Research/armada/internal/armada/configuration"
 )
 
 type LookoutDbMetricsProvider interface {
